Implement SSH client timeout in legacy fabric
This involved bumping the vendored SSH library forward 4 years. This
lead to having to provide an SSH host key validation method for all
client connections - I've chosen the one that ignores the host key,
because that's what we were doing up until now anyway.

The agent.Client code was only being used by its tests, and so I have
moved that set of code into the agent_test package.

The connection timeout is hard-coded to 30 seconds. This should only
account for the time it takes to establish the connection.
